refactor batch verifier to share with pending queue
